Use reflection bloq in chemistry decompositions.
